+++ finagle-memcached/src/test/java/com/twitter/finagle/memcached/integration/TestClient.java
@@ -1,18 +1,15 @@
 package com.twitter.finagle.memcached.integration;
 
-import java.util.ArrayList;
-
 import scala.Option;
-import scala.collection.JavaConversions;
 
 import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
 
 import com.twitter.finagle.Address;
-import com.twitter.finagle.Address$;
+import com.twitter.finagle.Addresses;
 import com.twitter.finagle.Memcached;
-import com.twitter.finagle.Name$;
+import com.twitter.finagle.Names;
 import com.twitter.finagle.Service;
 import com.twitter.finagle.memcached.JavaClient;
 import com.twitter.finagle.memcached.JavaClientBase;
@@ -20,7 +17,7 @@ import com.twitter.finagle.memcached.integration.external.TestMemcachedServer;
 import com.twitter.finagle.memcached.integration.external.TestMemcachedServer$;
 import com.twitter.finagle.memcached.protocol.Command;
 import com.twitter.finagle.memcached.protocol.Response;
-import com.twitter.io.Buf;
+import com.twitter.io.Bufs;
 import com.twitter.util.Await;
 
 import static org.junit.Assert.assertEquals;
@@ -39,18 +36,16 @@ public class TestClient {
    */
   @Test
   public void testGetAndSet() throws Exception {
-    Address addr = Address$.MODULE$.apply(server.get().address());
-    ArrayList<Address> addrs = new ArrayList<Address>();
-    addrs.add(addr);
+    Address addr = Addresses.newInetAddress(server.get().address());
 
     Service<Command, Response> service = Memcached.client()
           .connectionsPerEndpoint(1)
-          .newService(Name$.MODULE$.bound(JavaConversions.asScalaBuffer(addrs)), "memcached");
+          .newService(Names.bound(addr), "memcached");
 
     JavaClient client = JavaClientBase.newInstance(service);
     Await.ready(client.set("foo", "bar"));
 
-    Option<String> res = Buf.Utf8$.MODULE$.unapply(Await.result(client.get("foo")));
+    Option<String> res = Bufs.UTF_8.unapply(Await.result(client.get("foo")));
     assertEquals("bar", res.get());
   }
 }